FAR patch 01021.Mix Дата              : 24.10.2001 Сделал            : Valentin Skirdin Описание          : Mix Измененные файлы  : ctrlobj.cpp editor.cpp filepanels.cpp flupdate.cpp                     panel.cpp edit.cpp Новые файлы       : Удаленные файлы   : Состав            : 01021.Mix.txt                     ctrlobj.cpp.1021.diff       editor.cpp.1021.diff                     filepanels.cpp.1021.diff    flupdate.cpp.1021.diff                     panel.cpp.1021.diff         edit.cpp.1021.diff Основан на патче  : 1020 Дополнение        :
Редактор
~~~~~~~~
! обработка вставки имени файла и пути вынесена на уровень редактора, т.к.
  при выделеном блоке в несколько строк получаем лабуду.

! Корректно выставим последовательность отмены действий - DeleteBlock()
  сам умеет выставлять ундо

- Аналогичная ситуация с Ctrl-F - при выделенном не персистентном блоке
  сам блок не удаляется.


Панели
~~~~~~
  ! вместо "левая-правая" применим понятие "активная-пассивная" - так будет
    правильнее для ситуации "каталог создан не на той панели"
    Это при старте ФАРа и активизации панелей.

+++ far/edit.cpp
@@ -5,10 +5,14 @@ edit.cpp
 
 */
 
-/* Revision: 1.51 07.10.2001 $ */
+/* Revision: 1.52 24.10.2001 $ */
 
 /*
 Modify:
+  24.10.2001 SVS
+    - Ну разве так можно? Иван и ты, апологет персистентных блоков (?) не
+      заметил такой фигни? Для [Ctrl]ShiftEnter и Ctrl*Bracket блоки должны
+      удаляться только(!) если они не постоянные.
   07.10.2001 VVM
     + CTRL+SHIFT+ENTER вставляет файл с пассивной панели.
   14.09.2001 SKV
@@ -783,7 +787,8 @@ int Edit::ProcessKey(int Key)
             SelStart=PrevSelStart;
             SelEnd=PrevSelEnd;
           }
-          DeleteBlock();
+          if (!PersistentBlocks)
+            DeleteBlock();
           /* tran 03.07.2000 $ */
           /* $ 12.11.2000 KM
              Некоторые изменения вставки имени файла в строку ввода с маской.
@@ -832,7 +837,8 @@ int Edit::ProcessKey(int Key)
             SelStart=PrevSelStart;
             SelEnd=PrevSelEnd;
           }
-          DeleteBlock();
+          if (!PersistentBlocks)
+            DeleteBlock();
           /* tran 03.07.2000 $ */
           /* $ 10.07.2001 IS
              Вставляем строку с учетом текущей кодировки символов
